+++ actors/hierarchical-sca/Cargo.toml
@@ -28,7 +28,6 @@ serde = { version = "1.0.136", features = ["derive"] }
 anyhow = "1.0.56"
 fvm_ipld_blockstore = "0.1.0"
 fvm_ipld_encoding = "0.2.0"
-thiserror = "1.0.30"
 
 [dev-dependencies]
 fil_actors_runtime = { version = "8.0.0-alpha.1", path = "../runtime", features = ["test_utils", "sector-default"] }
