+++ actors/hierarchical-sca/src/checkpoint.rs
@@ -1,15 +1,10 @@
-use super::subnet::SubnetID;
-use crate::StorableMsg;
 use anyhow::anyhow;
 use cid::multihash::Code;
 use cid::multihash::MultihashDigest;
 use cid::Cid;
-use fil_actors_runtime::Array;
-use fvm_ipld_blockstore::MemoryBlockstore;
-use fvm_ipld_encoding::to_vec;
 use fvm_ipld_encoding::tuple::*;
-use fvm_ipld_encoding::CborStore;
-use fvm_ipld_encoding::{serde_bytes, Cbor};
+use fvm_ipld_encoding::{serde_bytes, to_vec, Cbor};
+use fvm_shared::address::SubnetID;
 use fvm_shared::bigint::bigint_ser;
 use fvm_shared::clock::ChainEpoch;
 use fvm_shared::econ::TokenAmount;
@@ -147,54 +142,6 @@ pub struct ChildCheck {
     pub checks: Vec<Cid>,
 }
 impl Cbor for ChildCheck {}
-
-#[derive(PartialEq, Eq, Clone, Debug, Default, Serialize_tuple, Deserialize_tuple)]
-pub struct CrossMsgs {
-    pub msgs: Vec<StorableMsg>,
-    pub metas: Vec<CrossMsgMeta>,
-}
-impl Cbor for CrossMsgs {}
-
-#[derive(PartialEq, Eq, Clone, Debug, Default, Serialize_tuple, Deserialize_tuple)]
-pub struct MetaTag {
-    pub msgs_cid: Cid,
-    pub meta_cid: Cid,
-}
-impl Cbor for MetaTag {}
-
-impl CrossMsgs {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub(crate) fn cid(&self) -> anyhow::Result<Cid> {
-        let store = MemoryBlockstore::new();
-        let mut msgs_array = Array::new(&store);
-        msgs_array.batch_set(self.msgs.clone())?;
-        let msgs_cid = msgs_array
-            .flush()
-            .map_err(|e| anyhow!("Failed to create empty messages array: {}", e))?;
-
-        let mut meta_array = Array::new(&store);
-        meta_array.batch_set(self.msgs.clone())?;
-        let meta_cid = meta_array
-            .flush()
-            .map_err(|e| anyhow!("Failed to create empty messages array: {}", e))?;
-
-        Ok(store.put_cbor(&MetaTag { msgs_cid: msgs_cid, meta_cid: meta_cid }, Code::Blake2b256)?)
-    }
-
-    pub(crate) fn add_metas(&mut self, metas: Vec<CrossMsgMeta>) -> anyhow::Result<()> {
-        for m in metas.iter() {
-            if self.metas.iter().any(|ms| ms == m) {
-                continue;
-            }
-            self.metas.push(m.clone());
-        }
-
-        Ok(())
-    }
-}
 
 /// CheckpointEpoch returns the epoch of the next checkpoint
 /// that needs to be signed
