+++ actors/hierarchical-sca/src/ext.rs
@@ -1,1 +1,3 @@
-pub mod subnet {}
+pub mod account {
+    pub const PUBKEY_ADDRESS_METHOD: u64 = 2;
+}
