+++ actors/hierarchical-sca/src/types.rs
@@ -1,12 +1,9 @@
 use super::checkpoint::{Checkpoint, CrossMsgMeta};
 use fil_actors_runtime::Array;
 use fvm_ipld_encoding::tuple::*;
-use fvm_ipld_encoding::RawBytes;
-use fvm_shared::address::Address;
 use fvm_shared::bigint::bigint_ser;
 use fvm_shared::clock::ChainEpoch;
 use fvm_shared::econ::TokenAmount;
-use fvm_shared::MethodNum;
 
 pub const CROSSMSG_AMT_BITWIDTH: u32 = 3;
 pub const DEFAULT_CHECKPOINT_PERIOD: ChainEpoch = 10;
@@ -35,21 +32,4 @@ pub struct CheckpointParams {
 #[derive(Debug, Serialize_tuple, Deserialize_tuple)]
 pub struct SubnetIDParam {
     pub id: String,
-}
-
-/// StorableMsg stores all the relevant information required
-/// to execute cross-messages.
-///
-/// We follow this approach because we can't directly store types.Message
-/// as we did in the actor's Go counter-part. Instead we just persist the
-/// information required to create the cross-messages and execute in the
-/// corresponding node implementation.
-#[derive(PartialEq, Eq, Clone, Debug, Serialize_tuple, Deserialize_tuple)]
-pub struct StorableMsg {
-    pub from: Address,
-    pub to: Address,
-    pub method: MethodNum,
-    pub params: RawBytes,
-    #[serde(with = "bigint_ser")]
-    pub value: TokenAmount,
 }
