+++ actors/hierarchical-sca/tests/sca_actor_test.rs
@@ -1,11 +1,11 @@
 use fil_actors_runtime::runtime::Runtime;
-use fvm_shared::address::Address;
+use fvm_shared::address::{Address, SubnetID};
 use fvm_shared::bigint::Zero;
 use fvm_shared::clock::ChainEpoch;
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::error::ExitCode;
 use hierarchical_sca::{
-    get_bottomup_msg, subnet, Actor as SCAActor, Checkpoint, CrossMsgMetaArray, State, SubnetID,
+    get_bottomup_msg, subnet, Actor as SCAActor, Checkpoint, CrossMsgMetaArray, State,
     DEFAULT_CHECKPOINT_PERIOD,
 };
 use std::str::FromStr;
